+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changes
 
+## Unreleased
+
+ * Generate token in LFS server instead pass original authentication data #105.
+
 ## 1.0.15-alhpa: GitLab 8.2 LFS storage layout support
 
  * Add support for GitLab 8.2 LFS storage layout #109.